Add FFB Strength to EEPROM

+++ ffb-pro.h
@@ -5,7 +5,7 @@
 #include <stdint.h>
 #include "ffb.h"
 
-extern float ffb_strength_ratio;
+extern uint8_t ffb_strength_ratio;
 
 // ----------------------------------
 // Microsoft Sidewinder Force Feedback Pro FFB structures
